requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anyio==3.7.1
app==0.0.1
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asttokens==2.4.0
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.1.0
Babel==2.13.0
backcall==0.2.0
backoff==2.2.1
backports.functools-lru-cache==1.6.5
bcrypt==4.0.1
beautifulsoup4==4.12.2
black==24.2.0
bleach==6.1.0
blinker==1.7.0
Brotli==1.1.0
cachetools==5.3.2
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.0
chroma-hnswlib==0.7.3
chromadb==0.4.14
click==8.1.7
codeboxapi==0.1.19
codeinterpreterapi==0.1.15
colorama==0.4.6
coloredlogs==15.0.1
comm==0.1.4
contourpy==1.2.0
cryptography==42.0.5
cycler==0.12.1
dataclasses-json==0.6.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distro==1.9.0
et-xmlfile==1.1.0
exceptiongroup==1.1.3
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.103.2
fastjsonschema==2.18.1
filelock==3.12.4
flatbuffers==23.5.26
fonttools==4.49.0
fqdn==1.5.1
frozenlist==1.4.0
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.40
google-api-core==2.17.1
google-auth==2.28.1
googleapis-common-protos==1.62.0
greenlet==3.0.3
grpcio==1.59.0
h11==0.14.0
hpack==4.0.0
httpcore==1.0.2
httptools==0.6.0
httpx==0.26.0
huggingface-hub==0.17.3
humanfriendly==10.0
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.1.0
iniconfig==2.0.0
ipykernel==6.25.2
ipython==8.16.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.2
jiter==0.5.0
joblib==1.4.2
jq==1.6.0
json5==0.9.14
jsonpatch==1.33
jsonpointer==2.4
jsonschema>=4.22.0,<5.0.0
jsonschema-specifications==2023.7.1
jupyter-events==0.9.0
jupyter-kernel-gateway==3.0.0
jupyter-lsp==2.2.0
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.13.0
jupyter_server_terminals==0.4.4
jupyterlab==4.0.7
jupyterlab-pygments==0.2.2
jupyterlab_server==2.25.0
kaleido==0.2.1
kiwisolver==1.4.5
langchain==0.1.11
langchain-community==0.0.27
langchain-core==0.1.52
langchain-openai==0.0.8
langchain-text-splitters==0.0.2
langchainhub==0.1.21
langgraph==0.0.26
langsmith==0.1.22
litellm==1.61.6
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.21.0
matplotlib==3.8.3
matplotlib-inline==0.1.6
mdurl==0.1.2
mediawikiapi==1.2
mistune==3.0.2
mkdocs==1.6.1
mkdocstrings==0.24.0
mkdocstrings-python
mkdocs-material[imaging]
mkdocs-minify-plugin
mkdocs-git-revision-date-localized-plugin      
monotonic==1.6
mpmath==1.3.0
multidict==6.0.4
mypy-extensions==1.0.0
nbclient==0.8.0
nbconvert==7.9.2
nbformat==5.9.2
nest-asyncio==1.5.8
networkx==3.4.2
notebook==7.0.5
notebook_shim==0.2.3
numexpr==2.9.0
numpy==1.26.0
onnxruntime==1.16.1
openai==1.61.0
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
orjson==3.9.15
overrides==7.4.0
packaging==23.2
pandas==2.1.1
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.12.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
plotly==5.20.0
pluggy==1.4.0
posthog==3.0.2
prometheus-client==0.17.1
prompt-toolkit==3.0.39
protobuf==4.25.2
psutil==5.9.5
psycopg2==2.9.10
ptyprocess==0.7.0
pulsar-client==3.3.0
pure-eval==0.2.2
pyarrow==14.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.6.3
pydantic-settings==2.2.1
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.16.1
pyOpenSSL==24.0.0
pyparsing==3.1.1
PyPika==0.48.9
PySocks==1.7.1
pytest==8.0.2
pytest-subtests==0.11.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytz==2023.3.post1
pyu2f==0.1.5
PyYAML==6.0.1
pyzmq==25.1.1
rdflib==7.0.0
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.6.0
rpds-py==0.10.6
rsa==4.9
scikit-learn==1.5.0
scipy==1.13.1
seaborn==0.13.2
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.5
SPARQLWrapper==2.0.0
SQLAlchemy==2.0.27
stack-data==0.6.2
starlette==0.27.0
streamlit==1.40.0
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
terminado==0.17.1
threadpoolctl==3.5.0
tiktoken==0.7.0
tinycss2==1.2.1
tokenize-rt==5.2.0
tokenizers==0.14.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
typer==0.9.0
types-python-dateutil==2.8.19.14
types-requests==2.31.0.20240125
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2023.3
tzlocal==5.2
uri-template==1.3.0
urllib3==2.0.6
uvicorn==0.23.2
uvloop==0.17.0
validators==0.22.0
watchfiles==0.21.0
wcwidth==0.2.8
webcolors==1.13
webencodings==0.5.1
websocket-client==1.6.4
websockets==12.0
wikibase-rest-api-client==0.1.3
yarl==1.9.2
zipp==3.17.0
